Exporter/OCAgent: Move TraceProtoUtils to exporter.
