+++ RELEASING.md
@@ -210,7 +210,6 @@ $ README_FILES=(
   contrib/log_correlation/log4j2/README.md
   contrib/log_correlation/stackdriver/README.md
   contrib/monitored_resource_util/README.md
-  contrib/opencensus_proto_util/README.md
   contrib/spring/README.md
   contrib/spring_sleuth_v1x/README.md
   contrib/zpages/README.md
