+++ all/build.gradle
@@ -15,7 +15,6 @@ def subprojects = [
         project(':opencensus-contrib-log-correlation-log4j2'),
         project(':opencensus-contrib-log-correlation-stackdriver'),
         project(':opencensus-contrib-monitored-resource-util'),
-        project(':opencensus-contrib-opencensus-proto-util'),
         project(':opencensus-contrib-spring'),
         project(':opencensus-contrib-spring-sleuth-v1x'),
         project(':opencensus-contrib-zpages'),
@@ -42,7 +41,6 @@ def subprojects_javadoc = [
         project(':opencensus-contrib-log-correlation-log4j2'),
         project(':opencensus-contrib-log-correlation-stackdriver'),
         project(':opencensus-contrib-monitored-resource-util'),
-        project(':opencensus-contrib-opencensus-proto-util'),
         project(':opencensus-contrib-spring'),
         project(':opencensus-contrib-spring-sleuth-v1x'),
         project(':opencensus-contrib-zpages'),
