+++ buildscripts/import-control.xml
@@ -134,12 +134,6 @@ General guidelines on imports:
     <subpackage name="monitoredresource.util">
       <allow pkg="io.opencensus.contrib.monitoredresource.util"/>
     </subpackage>
-    <subpackage name="opencensus.proto.util">
-      <allow pkg="com.google.protobuf"/>
-      <allow pkg="io.opencensus.contrib.opencensus.proto.util"/>
-      <allow pkg="io.opencensus.proto"/>
-      <allow pkg="io.opencensus.trace"/>
-    </subpackage>
   </subpackage>
   <subpackage name="exporter">
     <allow pkg="com.google.common"/>
@@ -175,8 +169,11 @@ General guidelines on imports:
         <allow pkg="org.apache.thrift"/>
       </subpackage>
       <subpackage name="ocagent">
+        <allow pkg="com.google.protobuf"/>
+        <allow pkg="io.opencensus.contrib.opencensus.proto.util"/>
         <allow pkg="io.opencensus.exporter.trace.ocagent"/>
         <allow pkg="io.opencensus.proto"/>
+        <allow pkg="io.opencensus.trace"/>
       </subpackage>
       <subpackage name="stackdriver">
         <allow pkg="com.google"/>
