--- contrib/opencensus_proto_util/README.md
@@ -1,24 +1,0 @@
-# OpenCensus Proto Util
-
-The *OpenCensus Proto Util for Java* is a collection of utilities for converting OpenCensus Java
-data models to OpenCensus proto data models.
-
-## Quickstart
-
-### Add the dependencies to your project
-
-For Maven add to your `pom.xml`:
-```xml
-<dependencies>
-  <dependency>
-    <groupId>io.opencensus</groupId>
-    <artifactId>opencensus-contrib-opencensus-proto-util</artifactId>
-    <version>0.17.0</version>
-  </dependency>
-</dependencies>
-```
-
-For Gradle add to your dependencies:
-```gradle
-compile 'io.opencensus:opencensus-contrib-opencensus-proto-util:0.17.0'
-```
