--- contrib/opencensus_proto_util/build.gradle
@@ -1,16 +1,0 @@
-description = 'OpenCensus Proto Util'
-
-apply plugin: 'java'
-
-[compileJava, compileTestJava].each() {
-    it.sourceCompatibility = 1.7
-    it.targetCompatibility = 1.7
-}
-
-dependencies {
-    compile project(':opencensus-api'),
-            libraries.opencensus_proto
-
-    signature "org.codehaus.mojo.signature:java17:1.0@signature"
-    signature "net.sf.androidscents.signature:android-api-level-14:4.0_r4@signature"
-}
